% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{drop_markers}
\alias{drop_markers}
\title{Drops markers by missing rate and/or minor allele frequency}
\usage{
drop_markers(mat, max_missing = 0.1, min_maf = 0.05)
}
\arguments{
\item{mat}{n x m marker matrix (numeric) for n individuals and markers}

\item{max_missing}{markers with mor than this fraction of missing values will be dropped}

\item{min_maf}{markers with a minor allele frequency lower than this will be dropped}
}
\value{
matrix with columns failing either \code{max_missing} or \code{min_maf} dropped
}
\description{
Drops markers by missing rate and/or minor allele frequency
}
