% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMAX_ANOVA.R
\name{EMMAX_ANOVA}
\alias{EMMAX_ANOVA}
\title{Title}
\usage{
EMMAX_ANOVA(
  formula,
  data,
  markers,
  genotypeID,
  cholL_Sigma_inv,
  mc.cores = RcppParallel::defaultNumThreads() - 1,
  verbose = T
)
}
\arguments{
\item{formula}{lm-style formula for the model to be fit for each marker.
The formula must include the special variable "X" which represents a marker.
\code{data} MUST NOT include "X" as a column, as this will be created in \code{data}
for each marker in the GWAS, replacing anything in that slot. Typically the formula
will include a main effect of each trait and an interaction between the marker and the trait.}

\item{data}{data.frame with at a minimum columns for trait and \code{genotypeID} to be used to match
up with \code{markers}. MUST NOT include a column titled "X".}

\item{markers}{matrix of marker genotypes with individuals as rows and rownames corresponding to the
\code{genotypeID} column of \code{data}}

\item{genotypeID}{column of \code{data} representing the genotype identifiers (rownames of \code{markers})}

\item{cholL_Sigma_inv}{lower-triangular matrix of the inverse of the Cholesky decomposition of the full variance-covariance
matrix of the data. Typically calculated with \code{make_cholL_Sigma_inv}}

\item{mc.cores}{Number of processor cores to use for parallel calculations}

\item{verbose}{Should progress across markers be reported?}
}
\value{
list of results with three elements:
    anovas: MSE, F-values and p-values for each term involving "X" for each marker
    beta_hats: coefficient estimates for each X term
    SEs: standard errors for each X term coefficient.
}
\description{
Title
}
