% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{fill_missing}
\alias{fill_missing}
\title{Fills in missing genotype values with the mean value for the marker}
\usage{
fill_missing(mat)
}
\arguments{
\item{mat}{n x m marker matrix (numeric) for n individuals and markers}
}
\value{
input matrix with NAs replaced with colMeans.
}
\description{
Note: this is a quick work-around, but ideally with dense markers more effective imputation should be used (e.g. Beagle)
}
