% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cholL_Sigma_inv.R
\name{make_cholL_Sigma_inv}
\alias{make_cholL_Sigma_inv}
\title{make_cholSigma}
\usage{
make_cholL_Sigma_inv(
  data,
  phenoID,
  rowID,
  columnID,
  covariances,
  sparse = TRUE
)
}
\arguments{
\item{data}{data.frame of observations}

\item{phenoID}{name of the phenotype column in \code{data}. This vector should not have any NAs (so run \code{subset(data,!is.na(data[[phenoID]]))} first).}

\item{rowID}{name of the column in \code{data} that would be the row-name of Y}

\item{columnID}{name of the column in \code{data} that would be the column-name of Y}

\item{covariances}{list of Row-Column pairs of covariance matrices.list(list(Row=K,Column=G),list(Row=I,Column=R)).
The Row matrices should have row names, and all elements of \code{data[[rowID]]} should be present in these row names.
If Row is missing, it is assumed to be the identity. The Column matrices should also have row names, and all elements
of \code{data[[columnID]]} should be present in the row names of these matrices.}

\item{sparse}{If TRUE and if the Cholesky decomposition of Sigma has >25% zeros, the output will be converted to a CSparseMatrix}
}
\value{
upper-triangle matrix with the inverse of the Cholesky decomposition.
}
\description{
Creates a full covariance matrix for a sparse matrix of data as the sum of
kronecker products of covariance matrix pairs and then computes the Cholesky decomposition, and finally the inverse of this upper-triangular matrix.
If Y is a (sparse) matrix
with distribution v(Y) ~ N(0,Sigma), where Sigma = \sum K_i \otimes G_i,
will return the Cholesky decomposition of Sigma[j,j] where j indexes the non-NA
elements of Y.
}
